+++ code/modules/mob/living/simple_animal/hostile/terror_spiders/queen_terror.dm
@@ -112,6 +112,8 @@
 			to_chat(T, "<span class='userdanger'>\The psychic backlash from the death of [src] overwhelms you! You feel the life start to drain out of you...</span>")
 			T.degenerate = TRUE
 		for(var/thing in GLOB.ts_spiderling_list)
+			if(!istype(thing, /mob/living/basic/spiderling/terror_spiderling))
+				continue
 			var/mob/living/basic/spiderling/terror_spiderling/T = thing
 			if(T.spider_myqueen && T.spider_myqueen == src)
 				qdel(T)
